requirements_completo.txt
fastapi
uvicorn
python-multipart
python-docx

